fix(auth): stabilize conditional passkey UI with ref-based callbacks
The useConditionalPasskeyUI hook recreated startConditionalUI every
render because callers passed inline onSuccess/refreshToken arrows.
Using refs for option callbacks decouples identity from the useCallback
dep array, so startConditionalUI stays stable across re-renders.

Also: extract hook to own module for testability, guard conditional UI
behind isAvailable check and on-prem flag, fix GenerateAuthOptionsResult
type to include hints field, remove redundant mountedRef initialization.
