+++ apps/web/src/components/auth/__tests__/useConditionalPasskeyUI.test.ts
@@ -1,15 +1,11 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { renderHook } from '@testing-library/react';
-import { useConditionalPasskeyUI } from '../PasskeyLoginButton';
+import { useConditionalPasskeyUI } from '../useConditionalPasskeyUI';
 
-// Mock all imports used by the hook
 vi.mock('@simplewebauthn/browser', () => ({
   startAuthentication: vi.fn(),
 }));
-vi.mock('@/components/ui/button', () => ({ Button: 'button' }));
-vi.mock('@/lib/utils', () => ({ cn: (...args: string[]) => args.join(' ') }));
 vi.mock('@/lib/utils/persist-csrf-token', () => ({ persistCsrfToken: vi.fn() }));
-vi.mock('@/hooks/useWebAuthnSupport', () => ({ useWebAuthnSupport: () => true }));
 vi.mock('@/stores/useAuthStore', () => ({
   useAuthStore: { getState: () => ({ setAuthFailedPermanently: vi.fn() }) },
 }));
@@ -18,9 +14,7 @@ vi.mock('@/lib/desktop-auth', () => ({
   handleDesktopAuthResponse: vi.fn().mockResolvedValue(false),
 }));
 vi.mock('sonner', () => ({ toast: { error: vi.fn(), success: vi.fn() } }));
-vi.mock('lucide-react', () => ({ Fingerprint: 'span', Loader2: 'span' }));
 
-// Stub conditional mediation as unavailable so the hook doesn't fire real fetches
 beforeEach(() => {
   vi.stubGlobal('PublicKeyCredential', {
     isConditionalMediationAvailable: () => Promise.resolve(false),
@@ -28,7 +22,7 @@ beforeEach(() => {
 });
 
 describe('useConditionalPasskeyUI', () => {
-  it('should return a stable startConditionalUI reference when options values are unchanged', () => {
+  it('given stable csrfToken and stable option refs, should return stable startConditionalUI', () => {
     const refreshToken = vi.fn();
     const onSuccess = vi.fn();
 
@@ -40,15 +34,12 @@ describe('useConditionalPasskeyUI', () => {
 
     const firstRef = result.current.startConditionalUI;
 
-    // Re-render with same csrfToken — options object is a new literal but values are identical
     rerender({ token: 'csrf-1' });
 
-    const secondRef = result.current.startConditionalUI;
-
-    expect(secondRef).toBe(firstRef);
+    expect(result.current.startConditionalUI).toBe(firstRef);
   });
 
-  it('should update startConditionalUI when csrfToken changes', () => {
+  it('given csrfToken changes, should return new startConditionalUI', () => {
     const refreshToken = vi.fn();
     const onSuccess = vi.fn();
 
@@ -62,8 +53,42 @@ describe('useConditionalPasskeyUI', () => {
 
     rerender({ token: 'csrf-2' });
 
-    const secondRef = result.current.startConditionalUI;
+    expect(result.current.startConditionalUI).not.toBe(firstRef);
+  });
 
-    expect(secondRef).not.toBe(firstRef);
+  it('given inline onSuccess arrow (new ref each render), should still return stable startConditionalUI', () => {
+    // This matches real usage in signin/page.tsx where onSuccess is an inline arrow
+    const { result, rerender } = renderHook(
+      ({ token }) =>
+        useConditionalPasskeyUI(token, {
+          refreshToken: async () => 'refreshed',
+          onSuccess: (url) => { window.location.href = url; },
+        }),
+      { initialProps: { token: 'csrf-1' } }
+    );
+
+    const firstRef = result.current.startConditionalUI;
+
+    // Re-render — options object and its callbacks are new refs
+    rerender({ token: 'csrf-1' });
+
+    expect(result.current.startConditionalUI).toBe(firstRef);
+  });
+
+  it('given inline refreshToken (new ref each render), should still return stable startConditionalUI', () => {
+    const { result, rerender } = renderHook(
+      ({ token }) =>
+        useConditionalPasskeyUI(token, {
+          refreshToken: async () => token,
+          onSuccess: vi.fn(),
+        }),
+      { initialProps: { token: 'csrf-1' } }
+    );
+
+    const firstRef = result.current.startConditionalUI;
+
+    rerender({ token: 'csrf-1' });
+
+    expect(result.current.startConditionalUI).toBe(firstRef);
   });
 });
